configuring judge model evaluator

+++ src/judge_evaluator.py
@@ -24,21 +24,26 @@ if __name__ == "__main__":
     # Argument parser
     parser = argparse.ArgumentParser()
     parser.add_argument("--empty_eval", type=str, required=True,
-                        help="Path to the evaluation dataset with empty assessments and feedback.")
+                        help="Path to the evaluation dataset with empty assessments and feedback.",
+                        default="caches/human-eval.json")
     parser.add_argument("--output_dir", type=str, required=True,
-                        help="Directory to store evaluation datasets for each model.")
-    parser.add_argument("--use_cache", type=bool, required=True,
-                        help="Decides whether or not to use cached judgements.")
-    parser.add_argument("--num_eval", type=bool, required=True,
-                        help="The number of times evaluations are made from which statistics are recorded.")
+                        help="Directory to store evaluation datasets for each model.",
+                        default="datasets")
+    parser.add_argument("--use_cache", type=str, required=True,
+                        help="If using cached evaluation files, provide the directory containing them.",
+                        default=True)
+    parser.add_argument("--num_eval", type=int, required=True,
+                        help="The number of times evaluations are made from which statistics are recorded.",
+                        default="2")
     args = parser.parse_args()
+
+    # Prepare LLM list
+    ollama_list = ["llama3.3:70b", "mistral-nemo:12b-instruct-2407-fp16"]
+    openai_list = ["gpt-4o-mini"]
 
     # Setup output directory
     output_dir = pathlib.Path(args.output_dir)
     interaction_path = output_dir / "interaction.json"
-    ollama_path = output_dir / "eval_ollama.json"
-    openai_path = output_dir / "eval_openai.json"
-    mistralnemo_path = output_dir / "eval_mistralnemo.json"
 
     # Read the evaluation dataset
     empty_eval_path = pathlib.Path(args.empty_eval)
@@ -49,9 +54,30 @@ if __name__ == "__main__":
     print("Loading empty evaluation dataset", flush=True)
     empty_eval_dataset = EvaluationDataset.model_validate_json(empty_eval_path.read_text())
 
-    print("Loading interactions dataset", flush=True)
-    interactions_dataset = extract_interaction_dataset(empty_eval_dataset)
-    interaction_path.write_text(interactions_dataset.model_dump_json(indent=4))
+    if not args.use_cache.exists():
+        print("No cache - Loading interactions dataset", flush=True)
+        interactions_dataset = extract_interaction_dataset(empty_eval_dataset)
+        interaction_path.write_text(interactions_dataset.model_dump_json(indent=4))
+
+
+    if args.use_cache.exists():
+        print("Caches evaluations being loaded", flush=True)
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
     print("Evaluating interactions as we speak", flush=True)
     # Ollama
